* README:           update properties in use         * src/com/fluendo/player/Cortado.java: (Cortado),         (Cortado.getParameterInfo), (Cortado.shutdown), (Cortado.init),         (AboutFrame.AboutFrame):         * src/com/fluendo/player/Status.java:           use a double for duration, is more intuitive than HMMSS           only mention audio backend details if the applet is configured to           play audio           show VM information in about box
git-svn-id: https://code.fluendo.com/flumotion/svn/cortado/trunk@3248 2b0047a9-a6d8-0310-accf-f7200b2a168c

+++ src/com/fluendo/player/Status.java
@@ -65,7 +65,7 @@ public class Status extends Component implements MouseListener,
 
     private double position = 0;
     private long time;
-    private long duration;
+    private double duration;
 
     private String speaker = "\0\0\0\0\0\357\0\0\357U\27"
             + "\36\0\0\0\0\357\357\0\0" + "\0\357U\30\0\0\0\357\0\357"
@@ -264,19 +264,19 @@ public class Status extends Component implements MouseListener,
         component.repaint();
     }
 
-    public void setTime(long seconds) {
+    public void setTime(double seconds) {
         if (clicked == NONE) {
             if (seconds < duration)
-                time = seconds;
+                time = (long) seconds;
             else
-                time = duration;
+                time = (long) duration;
 
             position = ((double) time) / duration;
             component.repaint();
         }
     }
 
-    public void setDuration(long seconds) {
+    public void setDuration(double seconds) {
         duration = seconds;
         component.repaint();
     }
